Task 15 - Spelling errors removed
Good afternoon Professor, 
In review of the document I had uploaded from my development box last night, I noticed this afternoon when looking over the files that there were a lot of spelling errors in the document.  I didn't change any of the content however I did re-upload the to correct the spelling errors and re-uploaded the file.  I am not sure what happened to the file as there were no spelling errors prior to the upload.  I hope that is alright.  You can choose to use whatever file you like.  Thank you so much and have a great day!